infinitelist: fix same key fetched
With the infiniteList when we reload the page that already been set.
We trying to fetch the same data with the same key. It not a good way
of programing and we have log error in the console. Now we concat the
data with an ES6 method that return the unique key of arrays. We do this
in order to fix that log error and to optimize the app.