InfiniteList: InfiniteList cound't find props
When we trying to mount the commonent after couple minute, we have some
trouble with infiniteList who's thrying to find props indefined. Now we
defined infined list if data have finish load. we to this in order to
fix this error and to mount correctly the component if data is exist.