util/getFetchedAt: cache desactivated
By default when an list item is not up to date to 10min,
it was removed from the list. but with infiniteList we have some trouble
with this. now desactivate this behavior in order to have infinteList
stable for now, we will change this behavior in nexted version of
infiniteList.