InfiniteList: query page by page
InfiniteList query all the pages because when we unmount and mount
the component we set the query page to 0 but the pagination should start
at 1, therefore postgres return an negative value that query all the
pages. Now we set this value to 1. we do this in order to fetch page
by page and in order to improuve render page rendering.