Button: stop event propagation
When we click on button from the new cliquablelist, we have propagation
of couple action from the button and the list who push to the history.
Now we stop the propagation event in order to have the right action
when we click on button of the list(createButton, DeleteButton,
Listbutton).
:construction: Notice that if you want to create an button with event
action on the list view you have to add an event.stopPropagation();