filter: changer the delays before fetching data
by default AOR make 500 ms before request fetch, we noticed that is to
speed for the users, now we set this value to 1000 ms. We do this in
order to have more time before fetching news data for now.
It will be powerfull if all the filters action are dissociated to the
list component.