InfiniteList: updating the component that already update
We have shouldComponentUpdate that already been defined by the List
component. InfiniteList extend to the list Component and it that all
List method by default. it make some trouble when we set filter search.
Now we deleted this method from infiniteList. we take the default method
from the List component.