input: AutoComplete, add condition before set the searchTest
On initialization, if the value was defined but the choices were not
fetched yet, the selector wasn't able to set the searchText, then once
the choices were fetched the component didn't try to update the
searchText.
So, we add a case to set the searchText in the input when
searchText is undefined && nextProps.input.value exists.