example: tests, fix e2e test for create view
1: we add data to the file data.js so when we try to create a view the
default id set wasn't the same than the previous version.
2: when we go to the view create by a direct link the global function
default values is called but not the one inside the component. We need
to refresh the view to get the default value. We did a little workaround
setting the value in the declaration in SimpleForm

NOTE: you may have an error like "component not clickable" it's because
you run the test on a too small screen when the bot can't reach the
button (he needs to scroll). run the test on a big screen will resole
that.